// src/app/core/models/contact.model.ts
export interface Contact {
  id?: number;
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber: string;
  address: string;
  city: string;
  state: string;
  country: string;
  postalCode: string;
  status?: 'active' | 'inactive' | 'blocked';
  createdDate?: Date;
  modifiedDate?: Date;
}